+++ packages/plugins/Filter/src/utils/arrayOnlyContains.ts
@@ -1,0 +1,2 @@
+export const arrayOnlyContains = (array: unknown[], value: unknown) =>
+  array.reduce((accumulator, current) => accumulator && current === value, true)