+++ packages/plugins/GeoLocation/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feature: Add parameter `renderType` to configuration, allowing configuration as IconMenu subcomponent.
 - Feature: Change the colour of tooltip to `primary` with its border being `primaryContrast` to be more visible.
+- Feature: Hide obstructive tooltip on small devices.
 
 ## 1.1.0